workflow/train_text_sql/deepspeed_trainer/requirements.txt
# DeepSpeed and distributed training
deepspeed>=0.12.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers and PEFT for QLoRA
transformers>=4.35.0
peft>=0.6.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Data processing
datasets>=2.14.0
tokenizers>=0.14.0
numpy>=1.24.0
pandas>=2.0.0

# Configuration and logging
pyyaml>=6.0
wandb>=0.15.0
tensorboard>=2.13.0
tqdm>=4.65.0

# Utilities
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
psutil>=5.9.0

# Optional: for specific model support
sentencepiece>=0.1.99
protobuf>=3.20.0

# Development tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 